Basic TypToP Linux PAM module and install script.
Main idea is that it "wraps" the pam_unix.so authentication.
If this fails, then pam_typtop is called to register a failed
password, and potentially override the decision.

If pam_unix succeeds, then pam_typtop is called to register
a successful password. The output of typtop is then ignored
(preventing typtop from locking the user out in case of
failure).